refactor(chat): Move provider-specific renderers into providers/ and add MCP tool rendering
Move all Claude Code-specific tool rendering logic from generic files
(toolRenderers.tsx, toolDetailRenderers.tsx, taskRenderers.tsx,
planModeRenderers.tsx) into providers/claudeRenderers.tsx, using the
existing renderMessage plugin hook — the same pattern Codex and ACP
providers already follow.

- Create providers/claudeRenderers.tsx with all Claude Code tool_use,
  tool_result, and agent_prompt rendering
- Add MCP tool_use rendering: parse mcp__server__tool names, show
  Wrench icon with humanized server/tool display name and input hint
- Move codexRenderers.tsx and opencodeRenderers.tsx into providers/
- Move Claude-specific test files into providers/
- Delete toolDetailRenderers.tsx and planModeRenderers.tsx (merged into
  claudeRenderers.tsx)
- Simplify messageRenderers.tsx generic dispatch (remove
  SPECIALIZED_TOOL_RENDERERS, dispatchToolUse, Claude-specific
  KIND_RENDERERS entries)
- Slim toolRenderers.tsx to only generic shared components
  (ToolUseLayout, ToolHeaderActions, ToolResultMessage, etc.)
