+++ frontend/src/components/chat/providers/codex.tsx
@@ -12,6 +12,9 @@ import * as workerRpc from '~/api/workerRpc'
 import { Icon } from '~/components/common/Icon'
 import { AgentProvider } from '~/generated/leapmux/v1/agent_pb'
 import * as styles from '../ChatView.css'
+import { CodexControlActions, CodexControlContent } from '../controls/CodexControlRequest'
+import { isNotificationThreadWrapper, isObject } from '../messageUtils'
+import { defaultModelId, effortItems, hasEfforts, modelDisplayName, modelItems, ModelSelect, optionGroup, optionGroupItems, optionLabel, PERMISSION_MODE_KEY, permissionModeGroup, permissionModeItems, RadioGroup } from '../settingsShared'
 import {
   codexAgentMessageRenderer,
   codexCollabAgentToolCallRenderer,
@@ -23,10 +26,7 @@ import {
   codexTurnCompletedRenderer,
   codexTurnPlanRenderer,
   codexWebSearchRenderer,
-} from '../codexRenderers'
-import { CodexControlActions, CodexControlContent } from '../controls/CodexControlRequest'
-import { isNotificationThreadWrapper, isObject } from '../messageUtils'
-import { defaultModelId, effortItems, hasEfforts, modelDisplayName, modelItems, ModelSelect, optionGroup, optionGroupItems, optionLabel, PERMISSION_MODE_KEY, permissionModeGroup, permissionModeItems, RadioGroup } from '../settingsShared'
+} from './codexRenderers'
 import { registerProvider } from './registry'
 
 /** Default model for Codex agents. */
