+++ frontend/src/components/chat/providers/toolSearchRenderer.test.tsx
@@ -1,5 +1,7 @@
+import type { MessageCategory } from '../messageClassification'
 import { render } from '@solidjs/testing-library'
 import { describe, expect, it, vi } from 'vitest'
+import { AgentProvider } from '~/generated/leapmux/v1/agent_pb'
 
 // eslint-disable-next-line no-control-regex -- ANSI escape detection requires matching control characters
 const ANSI_ESCAPE_RE = /\x1B\[[\d;]*m/
@@ -15,8 +17,8 @@ vi.mock('~/lib/renderMarkdown', () => ({
   renderMarkdown: (text: string) => text,
 }))
 
-const { renderMessageContent } = await import('./messageRenderers')
-type RenderContext = import('./messageRenderers').RenderContext
+const { renderMessageContent } = await import('../messageRenderers')
+type RenderContext = import('../messageRenderers').RenderContext
 
 /** Construct a ToolSearch tool_use assistant message. */
 function makeToolSearchMessage(input: Record<string, unknown> = {}) {
@@ -55,7 +57,9 @@ function makeToolSearchResult(
 /** Render a ToolSearch tool_use message and return its text content. */
 function renderToolUseText(input?: Record<string, unknown>, context?: RenderContext): string {
   const msg = makeToolSearchMessage(input)
-  const result = renderMessageContent(msg, 2 /* ASSISTANT */, context)
+  const toolUse = (msg.message.content as Array<Record<string, unknown>>)[0]
+  const category: MessageCategory = { kind: 'tool_use', toolName: 'ToolSearch', toolUse, content: msg.message.content as Array<Record<string, unknown>> }
+  const result = renderMessageContent(msg, 2 /* ASSISTANT */, context, category, AgentProvider.CLAUDE_CODE)
   const { container } = render(() => result)
   return container.textContent?.trim() ?? ''
 }
@@ -67,7 +71,8 @@ function renderToolResultContainer(
   context?: RenderContext,
 ): HTMLElement {
   const msg = makeToolSearchResult(toolRefs, toolUseResult)
-  const result = renderMessageContent(msg, 1 /* USER */, context)
+  const category: MessageCategory = { kind: 'tool_result' }
+  const result = renderMessageContent(msg, 1 /* USER */, context, category, AgentProvider.CLAUDE_CODE)
   const { container } = render(() => result)
   return container
 }
